stack-based variables are not allowed on async tasks.
When starting a async task we cannot use pointers to variables on the
stack, because when they will eventually be updated it will corrupt the
stack of some other task. Instead, move the variable in the task itself,
it will then become safe to alter at any moment.

Second thing fixed in this patch is the handling of the INFO parameter
for POTRF. This INFO should always be scaled to the correct diagonal
tile if an error is detected.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>